Generate tox.ini
The tox configuration format requires us to repeat ourselves quite a
bit, there are solutions to that such as substitutions that can
eliminate some of the repetition but still don't make it convenient to
express what is actually desired.

By generating the tox configuration and having the actual configuration
in Python code, we gain the desired power of expression.